Refactored the forms in ordinary previous address to be more reusable.
The ethos behind this was to group model classes and their mappings together,
since we are so often repeating the same mappings to bind the same case classes.

As a first pass at this I have implemented it on PreviousAddress on Ordinary.

Because I am trying to reuse mappings I have had to move the validations out of
the mappings and into the Forms that use those mappings.

E.g. the `PartialAddress.mapping` has a postcode field. Previously this was
validated to be a valid postcode. This validation now has to be applied each
place `PartialAddress.mapping` is used.

This required a lot of changes to the keys as for example of postcode:
- PreviousAddressFirstStep was using `previousAddress.postcode`
- PreviousAddressSelectStep was using `previousAddress.previousAddress.postcode`