refactor Contact/PostalVote forms to prevent circular references
In the PostalVote forms we need access to the contact forms, in the contact
forms we need access to the PostalVote forms.

To prevent this circular reference I have refactored the Contact forms to move
the contactMapping to a shared space, alongside the Contact case class, and the
PostalVote forms to move them into a shared space, alongside the PostalVote case
class.