Make test.conf accessible to jenkins
I think this fixes the bug. Tests pass here but that means little given they
have always passed locally.

The bug seems to be that the `test.conf` path was unreachable in jenkins.
We are now using `conf.resource` as it should look on the classpath of the
application and not require a full direct file path.