SessionHandling now uses SessionToken case class
I refactored the session handling to use a case class instead of a date time
string, providing the ability to add more functionality to the class.

As part of this I had to fix the JsonSerialisation, apparently we didn't notice
that our JsonSerialiser was serialising joda DateTimes with the Timezone
information but when deserialising it would always default to Timezone UTC.

To fix this I wrote a custom deserialiser that used the DateTime.parse method,
which recognises the Timezone information (the +1.00 stuff).