Session handling capable of handling old session tokens
Anyone who was previously using an old session token would when the new tokens
were brought in causing errors. This commit makes it so we can supress those
errors and refresh their session if that is the case.

I decided to go for the nuclear option (invalidate their session and wipe all
cookies) over the smarter try to handle the old style cookie because we aren't
live yet.