Cookies being set on domains
Cookies are now being set on the domain the request is coming in on, using the
'Host' header that is being proxied through by nginx. This ensures we don't
pass cookies on the assets domain.

The 'localhost' host name is being filtered out in ResultHandling.getDomain as
for some reason setting a cookie on the localhost domain causes the browser, or
chrome atleast, to not send the cookies on request. This causes the
`/register-to-vote/start` url to set a session cookie and the next page
`/register-to-vote/country-of-residence` looks for the cookie but the browser
hasn't sent it so it redirects the user back to the start page.