Versioning support for assets using the revision sha
This adds simple versioning to the assets so we can cache-bust on new deploys.
It does this by adding the current git sha to the asset path during the existing
remote asset path rewriting. I then added a stripping of the git sha in the
onRouteRequest method in DynamicGlobal.scala. This happens before routing so
the routes file can ignore the sha and route as normal.

On dev the default revision "abcdef1234567890abcdef1234567890abcdef12" will be
used.