Made the Mustachios possible to translate
I've added a trait that provides the messages to the Mustachios style of
mustaches. Also I've added an implicit function to make accessing the language
out of the request easier.

To use this take an exit or error mustache, e.g.

```
case class ServerError() extends GovukPage (
  "error/serverError",
  "Oops, we've done something wrong"
)
```

and mix in the trait MessagesForMustache and add an implicit Lang parameter:

```
case class ServerError() (
  implicit val lang:Lang
) extends GovukPage (
  "error/serverError",
  "Oops, we've done something wrong"
) with MessagesForMustache
```

This will grab the lang from the request implicitly and provide the correct
messages file to the mustache template. This means we can go ahead and use
`{{messages.foo}}` as in the other templates.