adding domain to google analytics
This is required for petes tests around the cross domain linking.
By using the `window.location.host` the cookie will be set on the host of the
current domain, preventing us from having to know what env the user is
currently looking at while also restricting the cookie so it doesn't get sent on
Asset calls.