Fixing ie8s preselecting of radio and checkboxes
IE8, unlike most browsers, returns an "" empty string when the checked attr is
not filled. Most browsers return null.

To fix this I have added a check to see that the value is not an "" empty string.

Problem is other browsers will return the content of checked, so if checked
exists and is not filled e.g. `<input type="checkbox" checked>` they will return
"". Thus this fix only works if we render proper checked="checked", which we do
at the moment. A better fix should be considered in the future.