Implemented local authority lookup using real api endpoint
Tests on the LocateService gss lookups

We had some issues over Guice being initialised with a set of bindings in a
previous test then not being destroyed before our test ran. In a previous commit
@januhlir encountered this problem and created FakeApplicationRefined.scala to
ensure that other tests were initialising the dependancies he needed for his
tests.

The problem with this then was that his LocateService was being injected when we
expected the actual LocateService to be used. To get our test to work for now we
have added this to the FakeApplicationRefined.scala:

    override def lookupGssCode(postcode: String): Option[String] = {
        Some("E09000030")
    }

This is not a good idea, we shouldn't be subjecting other tests to our mocking /
dependancies. The proper fix will be to ensure Guice gets destroyed at the end
of every test. Pivotal story https://www.pivotaltracker.com/story/show/74214040
has been created to implement the correct fix.