Fixed Nationality pages country labels
A previous commit added some html to the Nationality page through the messages
file. This had not been tested in a browser before merging so it wasn't noticed
that the `{{` escaped mustache tags were writing out the html as raw text.

This fixes that by switching the labels to `{{{` raw html mustache tags.