Fixing the Guice Container issues in Test
The GuiceContainer used to have issues in Tests where any dependancies
registered in a test at the top of the file would be available in tests at the
bottom of the file. This led to problems when writing tests as you couldn't
guarantee the state you were testing.

This fixes that by providing us a method to destroy the GuiceContainer and
hooking it in to the `Global.onStop` method. That way when a test is using a
FakeApplication to imitate a running app, at the test end the GuiceContainer
will be destroyed when the FakeApplication stops.

In order to get this code in I decided to duplicate it from [Guice-Utils]. A cop
out perhaps but given the projects spotty record for maintenance I think we are
more likely to make changes / develop if we can easily make those changes.

[Guice-Utils]: https://github.com/alphagov/gds-scala-common