Removed GuiceContainers mutability and switched DelegatingController to use Play style di
Having an object to access Guice is just a little scary since it isn't tied to
the Play application Global (which init and destroys it) and isn't tied to the
instances it's injected.

Solution, lets get rid of the GuiceContainer. Instead we'll new up the
GuiceContainer in the global and only use it from the global. To access it other
classes will either inject an Injector (it can be injected by Guice) or go through
play.api.Play.maybeApplication (which will fail if the app is not running).

Much safer.