Ensure Locate returns a gss if it can
This ensures if either of our datasets that contain gsscodes has a gsscode
for a particular postcode then the LocateService will return a gsscode. It will
do this by calling the Lookup Authority endpoint first then, if that has failed,
call the Lookup Address endpoint to try to get a gsscode from the address.

This should fix the bug introduced in [Fix manual address and gssCode submission lookup issue]

[Fix manual address and gssCode submission lookup issue]:https://github.gds/gds/ier-frontend/pull/522