Fix infinite recursion bug in FlowController
We had the possibility to have an infinite recursion bug in the Flow Controller
where the asking the current step for the next step could return the current step.

In the recusive function GoToNextIncompleteStep this would cause us to loop,
constantly asking the same step if it is complete and where to go next, constantly
answering yes and me.

This also in the Session Handling ensures that invalid sessions are completely
wiped rather than the current `Redirect(config.ordinaryStartUrl).withFreshSession()`.