Fixed dependency issue in new build server
It seems that signpost-http was requiring a version of org.apache.httpcomponents
httpclient that was behaving badly when we returned cookies on a redirect. This
presented as when the user clicked the Start button on the start page the akka
actor would return a Redirect with a session cookie attached but the response
would never reach the client connection.

The reason the old build worked was because the ivy.xml for the play_2.10-2.2.0.jar
bundled in the play.zip was forcing sbt to pull in fluentenlenium, which required
v4.2.1 of httpclient and this would override the misbehaving version of httpclient.

The fix ends up being just requiring as a dependency httpclient-4.2.1, forcing
it to override the broken version without having to pull in all of fluentenlenium.